backend/pyproject.toml
[project]
name = "cine-ai-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "qdrant-client>=1.12.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]
